synth_scripts/requirements.txt
py-aiger
matplotlib
networkx
numpy
pyverilog
seaborn
ipython
